Removed unused portion of code for template includes. Combined the template and environment _get_include methods into one.
